ENT-2653 Increase default max_stack_depth from default 2MB to 4MB
Changelog: Title

This change doubles the max_stack_depth from the default of 2MB to 4MB.

The postgresql.conf reccomends:

"The actual stack size limit enforced by the kernel (as set by ulimit-s or local
equvalent), less a safety margin of a megabyte or so. The safety margin is
needed because the stack depth is not checked in every routine in the server,
but only in key potentially-recursive routines such as expression evaluation.
The default setting is two megabytes (2MB), which is conservatively small and
unlikely to risk crashes."
https://www.postgresql.org/docs/10/static/runtime-config-resource.html#GUC-MAX-STACK-DEPTH

I found many references to settings as high as 10MB for this, but I found some
systems defaulting the stack size to 8MB, so 4MB seems to be a good improvement,
while remaining reasonably safe without dynamically calculating it.